feat: Integrate OAuth service and refactor authentication interfaces
- Added OAuthService to dependency injection in init_dependency.dart.
- Fixed import paths for AuthInterface in multiple files.
- Created a new AuthInterface with necessary methods for authentication.
- Implemented AuthInterfaceImpl to handle authentication logic, including Google login.
- Introduced CreateNewRideController for managing ride creation logic.
- Added rate ride functionality with RateRideController and associated models.
- Updated UI components to utilize new controllers and models for ride and rating features.
- Enhanced error handling and user feedback through SnackbarNotifier.
- Updated pubspec.yaml to include flutter_appauth dependency.
- Refactored ride-related models and interfaces for better clarity and functionality.

+++ lib/modules/ride&booking/model/rating.dart
@@ -1,6 +1,6 @@
 class Rating {
   final String userId;
-  final int score;
+  final double score;
 
   Rating({required this.userId, required this.score});
